requirements.txt
streamlit
ollama
pytesseract
easyocr
pdfplumber
requests
beautifulsoup4
numpy
pandas
pillow
asyncio
google